resource_sakuracloud_webaccel.go: structure_webaccel.goからスキーマ定義を移植
